Extract DNA genotype lookup helper
